pyproject.toml
[project]
name = "ci-starter"
version = "0.4.4"
description = "Kickstarts the semantic release pipeline for your Python project"
readme = "README.md"
authors = [
    { name = "fleetingbytes", email = "7075397+fleetingbytes@users.noreply.github.com" }
]
requires-python = ">=3.11"
license = "BSD-2-Clause"
keywords = ["semantic version", "semantic release", "ci/cd", "convetional commits", "python-semantic-release"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Unix Shell",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "click>=8.3",
    "gitpython>=3.1.45",
    "importlib-resources>=5.10 ; python_full_version < '3.12'",
    "platformdirs>=4.4.0",
    "ruamel-yaml>=0.18.15",
    "semver>=3.0.4",
    "tomli-w>=1.2.0",
]

[project.urls]
Homepage = "http://github.com/fleetingbytes/ci-starter"
Documentation = "http://github.com/fleetingbytes/ci-starter/blob/master/README.md"
Repository = "http://github.com/fleetingbytes/ci-starter.git"
Issues = "http://github.com/fleetingbytes/ci-starter/issues"
Changelog = "http://github.com/fleetingbytes/ci-starter/blob/master/CHANGELOG.md"

[project.scripts]
ci-start = "ci_starter.cli:cli"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
lint = [
    "ruff>=0.13.1",
]
test = [
    "pytest>=8.4.2",
]
